+++ lib/vcloud/core/query.rb
@@ -3,101 +3,59 @@ require 'csv'
 module Vcloud
   class Query
 
-    attr_reader :type
-    attr_reader :options
-
-    def initialize(type=nil, options={})
+    def initialize(type=nil, options={}, query_runner = Vcloud::QueryRunner.new)
       @type = type
       @options = options
       @options[:output_format] ||= 'tsv'
-      Fog.mock! if ENV['FOG_MOCK'] || options[:mock]
-      @fsi = Vcloud::Fog::ServiceInterface.new
-    end
-
-    def filter
-      options[:filter]
-    end
-
-    def output_format
-      options[:output_format]
-    end
-
-    def fields
-      options[:fields]
+      @query_runner = query_runner
     end
 
     def run()
-
-      puts "options:" if @options[:debug]
-      pp @options if @options[:debug]
-
       if @type.nil?
-        output_potential_query_types
+        output_available_query_types
       else
         output_query_results
       end
     end
 
+    # <b>DEPRECATED:</b> Please use <tt>Vcloud::QueryRunner.run</tt> instead.
     def get_all_results
-      results = []
-      (1..get_num_pages).each do |page|
-        results += get_results_page(page) || []
-      end
-      results
+      warn "[DEPRECATION] `Vcloud::Query::get_all_results` is deprecated.  Please use `Vcloud::QueryRunner.run` instead."
+      @query_runner.run(@type, @options)
     end
 
-    private
-
-    def get_num_pages
-      body = @fsi.get_execute_query(type=@type, @options)
-      last_page = body[:lastPage] || 1
-      raise "Invalid lastPage (#{last_page}) in query results" unless last_page.is_a? Integer
-      return last_page.to_i
-    end
-
-    def get_results_page(page)
-      raise "Must supply a page number" if page.nil?
-
-      begin
-        body = @fsi.get_execute_query(type=@type, @options.merge({:page=>page}))
-        pp body if @options[:debug]
-      rescue ::Fog::Compute::VcloudDirector::BadRequest, ::Fog::Compute::VcloudDirector::Forbidden => e
-        raise "Access denied: #{e.message}"
-      end
-
-      records = body.keys.detect {|key| key.to_s =~ /Record|Reference$/}
-      body[records] = [body[records]] if body[records].is_a?(Hash)
-      return nil if body[records].nil? || body[records].empty?
-      body[records]
-
-    end
-
+  private
     def output_query_results
-      results = get_all_results
+      results = @query_runner.run(@type, @options)
       output_header(results)
       output_results(results)
     end
 
-    def output_potential_query_types
+    def output_available_query_types
+      available_query_types = @query_runner.available_query_types
+      available_queries = collate_formats_for_types(available_query_types)
+      print_query_types(available_queries)
+    end
 
-      query_list = @fsi.get_execute_query
-      queries = {}
-      type_width = 0
-      query_list[:Link].select do |link|
-        link[:rel] == 'down'
-      end.map do |link|
-        href = Nokogiri::XML.fragment(link[:href])
-        query = CGI.parse(URI.parse(href.text).query)
-        [query['type'].first, query['format'].first]
-      end.each do |type, format|
-        queries[type] ||= []
+    def collate_formats_for_types(available_queries)
+      queries = Hash.new { |h, k| h[k]=[] }
+      available_queries.each do |type, format|
         queries[type] << format
-        type_width = [type_width, type.size].max
       end
+      queries
+    end
+
+    def print_query_types(queries)
+      type_width = longest_query_type(queries)
+
       queries.keys.sort.each do |type|
         puts "%-#{type_width}s %s" % [type, queries[type].sort.join(',')]
       end
+    end
 
+    def longest_query_type(queries)
+      return 0 if queries.keys.empty?
+      queries.keys.max_by{|key| key.length}.length
     end
 
     def output_header(results)
@@ -134,9 +92,7 @@ module Vcloud
       else
         raise "Unsupported output format #{@options[:output_format]}"
       end
-
     end
 
   end
 end
-