+++ spec/vcloud/core/edge_gateway_spec.rb
@@ -48,8 +48,9 @@ module Vcloud
           q_results = [
             { :name => 'edgegw-test-1', :href => "/#{@edgegw_id}" }
           ]
-          mock_query = double(:query, :get_all_results => q_results)
-          Vcloud::Query.should_receive(:new).with('edgeGateway', :filter => "name==edgegw-test-1").and_return(mock_query)
+          mock_query = double(:query_runner)
+          Vcloud::QueryRunner.should_receive(:new).and_return(mock_query)
+          mock_query.should_receive(:run).with('edgeGateway', :filter => "name==edgegw-test-1").and_return(q_results)
           @obj = EdgeGateway.get_by_name('edgegw-test-1')
           expect(@obj.class).to be(Vcloud::Core::EdgeGateway)
         end
@@ -58,16 +59,18 @@ module Vcloud
           q_results = [
             { :name => 'edgegw-test-1', :href => "/#{@edgegw_id}" }
           ]
-          mock_query = double(:query, :get_all_results => q_results)
-          Vcloud::Query.should_receive(:new).with('edgeGateway', :filter => "name==edgegw-test-1").and_return(mock_query)
+          mock_query = double(:query_runner)
+          Vcloud::QueryRunner.should_receive(:new).and_return(mock_query)
+          mock_query.should_receive(:run).with('edgeGateway', :filter => "name==edgegw-test-1").and_return(q_results)
           @obj = EdgeGateway.get_by_name('edgegw-test-1')
           expect(@obj.id) == @edgegw_id
         end
 
         it "should raise an error if no edgegw with that name exists" do
           q_results = [ ]
-          mock_query = double(:query, :get_all_results => q_results)
-          Vcloud::Query.should_receive(:new).with('edgeGateway', :filter => "name==edgegw-test-1").and_return(mock_query)
+          mock_query = double(:query_runner)
+          Vcloud::QueryRunner.should_receive(:new).and_return(mock_query)
+          mock_query.should_receive(:run).with('edgeGateway', :filter => "name==edgegw-test-1").and_return(q_results)
           expect{ EdgeGateway.get_by_name('edgegw-test-1') }.to raise_exception(RuntimeError, "edgeGateway edgegw-test-1 not found")
         end